chore: make dht tests deterministic
Because getClosestPeersOffline now only returns peers that are
closer than the caller, we have to sort the test dhts by closeness
to query targets before using them.
