+++ packages/kad-dht/test/kad-dht.spec.ts
@@ -1,17 +1,14 @@
 /* eslint-env mocha */
 /* eslint max-nested-callbacks: ["error", 8] */
 
-import { peerIdFromMultihash } from '@libp2p/peer-id'
 import { Libp2pRecord } from '@libp2p/record'
 import { expect } from 'aegir/chai'
-import delay from 'delay'
 import all from 'it-all'
 import drain from 'it-drain'
 import filter from 'it-filter'
 import last from 'it-last'
 import map from 'it-map'
 import { pipe } from 'it-pipe'
-import * as Digest from 'multiformats/hashes/digest'
 import sinon from 'sinon'
 import { equals as uint8ArrayEquals } from 'uint8arrays/equals'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
@@ -20,15 +17,12 @@ import { EventTypes, MessageType } from '../src/index.js'
 import { peerResponseEvent } from '../src/query/events.js'
 import * as kadUtils from '../src/utils.js'
 import { createPeerIdsWithPrivateKey } from './utils/create-peer-id.js'
-import { createValues } from './utils/create-values.js'
 import { countDiffPeers } from './utils/index.js'
 import { sortClosestPeers, sortDHTs } from './utils/sort-closest-peers.js'
 import { TestDHT } from './utils/test-dht.js'
 import type { PeerIdWithPrivateKey } from './utils/create-peer-id.js'
 import type { FinalPeerEvent, QueryEvent, ValueEvent } from '../src/index.js'
 import type { KadDHT } from '../src/kad-dht.js'
-import type { PeerId } from '@libp2p/interface'
-import type { CID } from 'multiformats/cid'
 
 async function findEvent (events: AsyncIterable<QueryEvent>, name: 'FINAL_PEER'): Promise<FinalPeerEvent>
 async function findEvent (events: AsyncIterable<QueryEvent>, name: 'VALUE'): Promise<ValueEvent>
@@ -51,7 +45,6 @@ async function findEvent (events: AsyncIterable<QueryEvent>, name: string): Prom
 
 describe('KadDHT', () => {
   let peerIds: PeerIdWithPrivateKey[]
-  let values: Array<{ cid: CID, value: Uint8Array }>
   let tdht: TestDHT
 
   beforeEach(() => {
@@ -64,14 +57,7 @@ describe('KadDHT', () => {
 
   before(async function () {
     this.timeout(10 * 1000)
-
-    const res = await Promise.all([
-      createPeerIdsWithPrivateKey(3),
-      createValues(20)
-    ])
-
-    peerIds = res[0]
-    values = res[1]
+    peerIds = await createPeerIdsWithPrivateKey(3)
   })
 
   afterEach(() => {
@@ -411,227 +397,6 @@ describe('KadDHT', () => {
 
       const res = await last(dht.get(key))
       expect(res).to.have.property('value').that.equalBytes(value)
-    })
-  })
-
-  describe('content routing', () => {
-    it('provides', async function () {
-      this.timeout(20 * 1000)
-
-      const dhts = await Promise.all([
-        tdht.spawn(),
-        tdht.spawn(),
-        tdht.spawn(),
-        tdht.spawn()
-      ])
-
-      const ids = dhts.map((d) => d.components.peerId)
-      const idsB58 = ids.map(id => id.toString())
-      sinon.spy(dhts[3].network, 'sendMessage')
-
-      // connect peers
-      await Promise.all([
-        tdht.connect(dhts[0], dhts[1]),
-        tdht.connect(dhts[1], dhts[2]),
-        tdht.connect(dhts[2], dhts[3])
-      ])
-
-      // provide values
-      await Promise.all(values.map(async (value) => { await drain(dhts[3].provide(value.cid)) }))
-
-      // Expect an ADD_PROVIDER message to be sent to each peer for each value
-      const fn = dhts[3].network.sendMessage
-      const valuesBuffs = values.map(v => v.cid.multihash.bytes)
-      // @ts-expect-error fn is a spy
-      const calls = fn.getCalls().map(c => c.args)
-
-      for (const [peerId, msg] of calls) {
-        expect(idsB58).includes(peerId.toString())
-        expect(msg.type).equals(MessageType.ADD_PROVIDER)
-        expect(valuesBuffs).includes(msg.key)
-        expect(msg.providers.length).equals(1)
-        expect(peerIdFromMultihash(Digest.decode(msg.providers[0].id)).toString()).equals(idsB58[3])
-      }
-
-      // Expect each DHT to find the provider of each value
-      let n = 0
-      for (const v of values) {
-        n = (n + 1) % 3
-
-        const events = await all(dhts[n].findProviders(v.cid))
-        const provs = Object.values(events.reduce<Record<string, PeerId>>((acc, curr) => {
-          if (curr.name === 'PEER_RESPONSE') {
-            curr.providers.forEach(peer => {
-              acc[peer.id.toString()] = peer.id
-            })
-          }
-
-          return acc
-        }, {}))
-
-        expect(provs).to.have.length(1)
-        expect(provs[0].toString()).to.equal(ids[3].toString())
-      }
-    })
-
-    it('provides if in server mode', async function () {
-      const dhts = await Promise.all([
-        tdht.spawn(),
-        tdht.spawn(),
-        tdht.spawn(),
-        tdht.spawn()
-      ])
-
-      // connect peers
-      await Promise.all([
-        tdht.connect(dhts[0], dhts[1]),
-        tdht.connect(dhts[1], dhts[2]),
-        tdht.connect(dhts[2], dhts[3])
-      ])
-
-      const sendMessageSpy = sinon.spy(dhts[0].network, 'sendMessage')
-
-      await dhts[0].setMode('server')
-
-      await drain(dhts[0].provide(values[0].cid))
-
-      expect(sendMessageSpy.called).to.be.true()
-    })
-
-    it('find providers', async function () {
-      this.timeout(20 * 1000)
-
-      const val = values[0]
-
-      const dhts = await Promise.all([
-        tdht.spawn(),
-        tdht.spawn(),
-        tdht.spawn()
-      ])
-
-      // Connect
-      await Promise.all([
-        tdht.connect(dhts[0], dhts[1]),
-        tdht.connect(dhts[1], dhts[2])
-      ])
-
-      await Promise.all(dhts.map(async (dht) => { await drain(dht.provide(val.cid)) }))
-
-      const events = await all(dhts[0].findProviders(val.cid))
-
-      // find providers find all the 3 providers
-      const provs = Object.values(events.reduce<Record<string, PeerId>>((acc, curr) => {
-        if (curr.name === 'PEER_RESPONSE') {
-          curr.providers.forEach(peer => {
-            acc[peer.id.toString()] = peer.id
-          })
-        }
-
-        return acc
-      }, {}))
-      expect(provs).to.have.length(3)
-    })
-
-    it('find providers from client', async function () {
-      this.timeout(20 * 1000)
-
-      const val = values[0]
-
-      const dhts = await Promise.all([
-        tdht.spawn(),
-        tdht.spawn(),
-        tdht.spawn()
-      ])
-      const clientDHT = await tdht.spawn({ clientMode: true })
-
-      // Connect
-      await Promise.all([
-        tdht.connect(clientDHT, dhts[0]),
-        tdht.connect(dhts[0], dhts[1]),
-        tdht.connect(dhts[1], dhts[2])
-      ])
-
-      await Promise.all(dhts.map(async (dht) => { await drain(dht.provide(val.cid)) }))
-
-      const events = await all(dhts[0].findProviders(val.cid))
-
-      // find providers find all the 3 providers
-      const provs = Object.values(events.reduce<Record<string, PeerId>>((acc, curr) => {
-        if (curr.name === 'PEER_RESPONSE') {
-          curr.providers.forEach(peer => {
-            acc[peer.id.toString()] = peer.id
-          })
-        }
-
-        return acc
-      }, {}))
-      expect(provs).to.have.length(3)
-    })
-
-    it('find client provider', async function () {
-      this.timeout(20 * 1000)
-
-      const val = values[0]
-
-      const dhts = await Promise.all([
-        tdht.spawn(),
-        tdht.spawn()
-      ])
-      const clientDHT = await tdht.spawn({ clientMode: true })
-
-      // Connect
-      await Promise.all([
-        tdht.connect(clientDHT, dhts[0]),
-        tdht.connect(dhts[0], dhts[1])
-      ])
-
-      await drain(clientDHT.provide(val.cid))
-
-      await delay(1e3)
-
-      const events = await all(dhts[1].findProviders(val.cid))
-
-      // find providers find the client provider
-      const provs = Object.values(events.reduce<Record<string, PeerId>>((acc, curr) => {
-        if (curr.name === 'PEER_RESPONSE') {
-          curr.providers.forEach(peer => {
-            acc[peer.id.toString()] = peer.id
-          })
-        }
-
-        return acc
-      }, {}))
-      expect(provs).to.have.length(1)
-    })
-
-    it('find one provider locally', async function () {
-      this.timeout(20 * 1000)
-      const val = values[0]
-
-      const dht = await tdht.spawn()
-
-      sinon.stub(dht.components.peerStore, 'get').withArgs(dht.components.peerId)
-        .resolves({
-          id: dht.components.peerId,
-          addresses: [],
-          protocols: [],
-          tags: new Map(),
-          metadata: new Map()
-        })
-      sinon.stub(dht.providers, 'getProviders').resolves([dht.components.peerId])
-
-      // Find provider
-      const events = await all(dht.findProviders(val.cid))
-      const provs = Object.values(events.reduce<Record<string, PeerId>>((acc, curr) => {
-        if (curr.name === 'PEER_RESPONSE') {
-          curr.providers.forEach(peer => {
-            acc[peer.id.toString()] = peer.id
-          })
-        }
-
-        return acc
-      }, {}))
-      expect(provs).to.have.length(1)
     })
   })
